backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
openpyxl
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.1
google-generativeai==0.8.5
google-genai
